requirements.txt
fastapi==0.85.0
uvicorn==0.18.2
openai==0.27.0
langchain==0.0.101
langchain-openai
langchain_community
sentence-transformers==2.2.2
langchain-chroma>=0.1.2
chromadb==0.3.21
pytest==7.1.2
python-dotenv==0.21.0
docx2txt